backend/requirements.txt
flask
flask-socketio
flask-sqlalchemy
flask-migrate
flask-cors
flask-bcrypt
flask-jwt-extended
psycopg2-binary
python-dotenv
cloudinary
eventlet
